ants_adk/requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
rank-bm25>=0.2.2
